remove unnecessary deps. refactoring (spelling), licsense, etc
