+++ math-scala/pom.xml
@@ -140,11 +140,6 @@
       <version>3.3.0</version>
     </dependency>
 
-    <dependency>
-      <groupId>com.github.haifengl</groupId>
-      <artifactId>smile-core</artifactId>
-      <version>1.1.0</version>
-    </dependency>
 
     <dependency>
       <groupId>com.github.haifengl</groupId>
