--- math-scala/src/main/scala/org/apache/mahout/visiualization/mlpot.scala
@@ -1,8 +1,0 @@
-package org.apache.mahout.visiualization
-
-/**
-  * Created by andy on 4/27/16.
-  */
-trait mlpot extends {
-
-}
