--- math-scala/src/main/scala/org/apache/mahout/visiualization/mplot3d.scala
@@ -1,8 +1,0 @@
-package org.apache.mahout.visiualization
-
-/**
-  * Created by andy on 4/27/16.
-  */
-class mplot3d {
-
-}
