+++ pom.xml
@@ -469,29 +469,6 @@
       </dependency>
 
       <dependency>
-        <groupId>com.github.haifengl</groupId>
-        <artifactId>smile-core</artifactId>
-        <version>1.1.0</version>
-      </dependency>
-      <dependency>
-        <groupId>com.github.haifengl</groupId>
-        <artifactId>smile-math</artifactId>
-        <version>1.1.0</version>
-      </dependency>
-
-      <dependency>
-        <groupId>com.github.haifengl</groupId>
-        <artifactId>smile-plot</artifactId>
-        <version>1.1.0</version>
-      </dependency>
-
-      <dependency>
-        <groupId>org.swinglabs</groupId>
-        <artifactId>swingx</artifactId>
-        <version>1.6.1</version>
-      </dependency>
-
-      <dependency>
         <groupId>org.apache.hbase</groupId>
         <artifactId>hbase-client</artifactId>
         <version>${hbase.version}</version>
